refactor: modularize type system into domain-specific modules
