+++ src/components/layout/TimelineContainer.tsx
@@ -1,6 +1,6 @@
 // src/components/timeline/TimelineContainer.tsx
 
-import type { TimelineContainerProps } from '~/types';
+import type { TimelineContainerProps } from '~/types/layout';
 export function TimelineContainer({ children, className = '' }: TimelineContainerProps) 
 {
   return (
