+++ src/components/navigation/NavigationArrow.tsx
@@ -1,6 +1,6 @@
 // src/components/ui/cards/NavigationArrow.tsx
 
-import type { NavigationArrowProps } from '~/types';
+import type { NavigationArrowProps } from '~/types/navigation';
 
 export function NavigationArrow({ direction, onClick, disabled }: NavigationArrowProps) 
 {
