+++ src/components/navigation/SectionNavButton.tsx
@@ -1,7 +1,7 @@
 // src/components/ui/SectionNavButton.tsx
 
 import Link from 'next/link';
-import type { SectionNavButtonProps } from '~/types';
+import type { SectionNavButtonProps } from '~/types/navigation';
 
 export function SectionNavButton({ href, children, className = '' }: SectionNavButtonProps) 
 {
